+++ lib/mp_disp.m
@@ -6,6 +6,9 @@ function mp_disp(obj)
 %   mp_disp(something_to_be_displayed);
 %   mp.logger.manager('clear');
 %
+% Optionally redirects the output of ``disp()`` to a file via an mp.logger
+% object, or elsewhere via a custom mp.logger subclass.
+%
 % See also mp.logger.manager, mp.logger.
 
 %   MP-Test